Refactor issue/view/view.go to introduce IssuePrint struct
The entire tty printing of this command was tested through fragile blocks
of strings that made any sort of testing or refactoring difficult, as I
found in the previous commit. I found myself iterating back and forth
with the tests and the code in order to make sure my expected output was
just right to match what the code was doing.

To simplify this, I've introduced the IssuePrint struct, which contains
methods for each of the individual pieces of an issue that will be
printed. This not only makes understanding what the terminal output will
look like easier, but allows us to both test these functions individually
(which I'll do in a forthcoming commit) and mock them for simpler testing
of the larger command.

In this commit, I have not added any functions beyond changing them to
work with the new struct. All tests are green. In addition, there were a
few minor bugfixes that this new format found, such as a nil pointer issue
in the milestone printing section, that have been fixed as well.